Refactor the steering controller.
This commit introduces a new module, `Ada_Steering_Controller` that
encapsulates the state and functionality of the control solution for
steering the RC car. This encapsulation will facilitate exploration of
alternative control solutions without requiring significant rewriting of
the `Steering_Control` task body.

The refactoring has not been tested, as I don't have the hardware needed
to execute the code.